feat(products): 전체 상품 랭킹 어댑터 수정 및 추가
- 전체 랭킹 화면의 어댑터 ProductAdapter에 상품 아이템 클릭 이벤트 로직 추가, 액티비티 코드에도 추가
- 홈 화면의 전체 랭킹 조회를 위한 어댑터 HomeRankingAdapter 생성 및 설정

+++ app/src/main/java/com/nuda/nudaclient/data/remote/dto/products/ProductsGetAllRankingResponse.kt
@@ -8,7 +8,7 @@ data class ProductsGetAllRankingResponse(
     val hasNext: Boolean
 ) {
     data class NextCursor(
-        val sortValue: Int,
+        val sortValue: Double,
         val id: Int
     )
 }
